+++ NEWS.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## New features
+
+* Enable filtering by the achievment level of the qi([#89](https://github.com/SKDE-Felles/qmongr/pull/89))
+
 # qmongr 0.4.0
 
 * Add qi overview ([#84](https://github.com/SKDE-Felles/qmongr/pull/84))